pyproject.toml
[project]
name = "bookly-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "asgiref>=3.11.1",
    "asyncpg>=0.31.0",
    "bcrypt==3.2.2",
    "celery>=5.6.2",
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.0",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "pwdlib>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "redis>=7.1.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
